fix(parser): shrink data-store-transport.ts under its module-size budget
CI's module-size gate failed PR #3606 (fix-quantity-table-duplicate-qset-instance,
head 5fb0638bd): packages/parser/src/data-store-transport.ts measured 531
lines against its 530-line budget, both on the branch itself and after
merging main (main's copy is 530, unchanged).

Extract the CompactEntityIndex transport (its column interface plus
toColumns/fromColumns) into a sibling packages/parser/src/compact-entity-index-transport.ts.
It only handled CompactEntityIndex serialization, a self-contained
concern distinct from the rest of the file's IfcDataStore transport, so
splitting it out doesn't need touching any other logic. Names and
behaviour of the two functions and CompactEntityIndexColumns are
unchanged; data-store-transport.ts re-exports the type and imports the
functions so no caller changes. data-store-transport.ts is now 491
lines (39 lines of headroom against 530); the new file is 56 lines and
carries no allowlist row, per house rules ("shrink or split instead of
raising the budget").
